Update .po translation/localization files
Two key changes were made in this commit. Firstly, the timestamp for the creation of the 'django.po' file in the 'sw' locale was updated. Next, new validation messages were added to the same file responding to specific scenarios in community appointments and facility-based participations.

A new 'django.po' file was also created in the 'lg' locale as per the project's requirement to support this language. The file contains several translations related to appointment scheduling and patient details, setting up the groundwork for further localization work in this language. This will help in expanding the application's language support and usability.